fix(lens): sibling ships only wait-then-SIGKILL
Delete the unused TERM→KILL ladder the sibling no longer calls.
Move argv dispatch to bindPidWatchdog.cli.ts. Share the bind-pid
poll cadence with waitForShutdown.

- lowy L2 / hickey H4 — dead reapUncooperative on the daemon side
- hickey H5 — library+argv in one module
- hickey H8 — two 2000ms poll constants

Raised by the lowy ∥ hickey lens review. Not pushed or merged.
